+++ product/dashboard/src/systemTest/resources/GebConfig.groovy
@@ -1,7 +1,6 @@
 import org.openqa.selenium.chrome.ChromeDriver
 import org.openqa.selenium.chrome.ChromeOptions
 import org.openqa.selenium.firefox.FirefoxDriver
-import org.openqa.selenium.phantomjs.PhantomJSDriver
 
 environments {
 
@@ -22,12 +21,6 @@ environments {
     firefox {
         driver = {
             new FirefoxDriver()
-        }
-    }
-
-    phantomjs {
-        driver = {
-            new PhantomJSDriver()
         }
     }
 }